requirements.txt
MySQL-python==1.2.3
SQLAlchemy==0.7.8
benchmark==0.1.5
decorator==3.3.3
nose==1.2.0
psycopg2==2.4.5
python-faker==0.2.4
wsgiref==0.1.2
